+++ components/partnerships/start-page/FeaturesSection.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const FeaturesSection = ({ textContent }) => {
+  return (
+    <section className="overflow-hidden bg-gray-1">
+      <div></div>
+    </section>
+  );
+};
+
+export default FeaturesSection;
